+++ tests/verbs/test_compute_communities.py
@@ -5,7 +5,7 @@ from datashaper import NoopVerbCallbacks
 
 from graphrag.config.create_graphrag_config import create_graphrag_config
 from graphrag.index.run.utils import create_run_context
-from graphrag.index.workflows.v1.compute_communities import run_workflow
+from graphrag.index.workflows.compute_communities import run_workflow
 
 from .util import (
     compare_outputs,
